+++ ember-cli-build.js
@@ -3,11 +3,11 @@ const EmberApp = require('ember-cli/lib/broccoli/ember-app');
 
 module.exports = function(defaults) {
   const app = new EmberApp(defaults, {
-    sassOptions: {},
+    'sassOptions': {},
 
     'autoprefixer': {
       browsers: ['last 3 versions'],
-    },
+    }
   });
 
   // Use `app.import` to add additional libraries to the generated
