+++ package.json
@@ -35,6 +35,7 @@
     "ember-export-application-global": "^1.0.5",
     "ember-load-initializers": "^0.6.0",
     "ember-resolver": "^2.0.3",
+    "ember-route-action-helper": "^2.0.2",
     "ember-source": "~2.12.0",
     "loader.js": "^4.2.3"
   },
