--- tests/unit/controllers/quiz-test.js
@@ -1,12 +1,0 @@
-import { moduleFor, test } from 'ember-qunit';
-
-moduleFor('controller:quiz', 'Unit | Controller | quiz', {
-  // Specify the other units that are required for this test.
-  // needs: ['controller:foo']
-});
-
-// Replace this with your real tests.
-test('it exists', function(assert) {
-  let controller = this.subject();
-  assert.ok(controller);
-});
